Added epoch_length compensation and moved text related functions to text_dataloader
