+++ bittensor/dataloaders/dataloader.py
@@ -1,13 +1,14 @@
 import argparse
 import bittensor
-import random
 import requests
+import random
 from munch import Munch
 
-from loguru import logger
 from requests.adapters import HTTPAdapter
 from requests.packages.urllib3.util.retry import Retry
 from torch.utils.data.dataloader import DataLoader
+from torch.utils.data import Subset
+
 
 
 class BittensorDataLoader():
@@ -16,7 +17,10 @@ class BittensorDataLoader():
         # TODO (shibshib): Find a proper way to set this as config instead of hardcoding it.
         # More dataset hashes can be added as we add directories for other modalities.
         self.genesis_text_dataset_hash = "QmXwfPoh2QFYqC6cYcW8kzyd9ruFfhnUi2kVBkdhawjUzj"
+
+        # Used to retrieve directory contentx
         self.dag_get = 'https://ipfs.infura.io:5001/api/v0/dag/get'
+        # Used to retrieve file contents
         self.file_cat = 'https://ipfs.infura.io:5001/api/v0/cat'
         
     @staticmethod   
@@ -39,7 +43,6 @@ class BittensorDataLoader():
     @staticmethod   
     def check_config(config: Munch):
         pass
-    
 
     @staticmethod
     def requests_retry_session(
@@ -93,56 +96,40 @@ class BittensorDataLoader():
         
         return directory
     
-    def retrieve_text_file(self, file_hash: str):
-        """Connects to Infura IPFS gateway and retrieves the contents of 
-        a genesis text file.
-
-        Returns:
-            str: The contents of the file.
-        """
-        session = requests.Session()
-        params = (('arg', file_hash),)
-        session.params.update(params)
-        directory = None
-
-        response = BittensorDataLoader.requests_retry_session(session=session).post(self.file_cat)
-
-        if response.status_code == 200:
-            directory = response
-        
-        return directory       
-
-    def retrieve_random_text_data(self):
-        """Connects to Infura IPFS gateway and retrieves the directory of genesis datasets.
-        
-        Returns:
-            string: Contents of the text file. 
-        """
-        try:
-            logger.info("Retrieving a dataset file from the IPFS gateway...")
-            directory = self.retrieve_directory(self.genesis_text_dataset_hash)
-
-            # Pick a random dataset file and return its contents
-            if directory and 'links' in directory.keys():
-                random_dataset_file = random.choice(directory['links'])
-                filename = random_dataset_file['Name']
-                # Make sure the file we chose satisfies our maximum file size requirement
-                if random_dataset_file['Size'] <= self.config.dataloader.max_file_size:
-                    random_dataset_file_hash = random_dataset_file['Cid']['/']
-                    file_contents = self.retrieve_text_file(random_dataset_file_hash)
-                    logger.info("Retrieved {} as training data...".format(filename))
-                    return filename, file_contents.text.split()
-
-            logger.error("It appears the directory is empty... Restart your miner to try again.")
-            return None
-        except Exception as ex:
-            logger.error("Ran into exception when trying to retrieve dataset from IPFS: {}".format(ex))
-
-        return None
     
-    def dataloader(self):
-        # Convert into a torch dataloader
-        return DataLoader(self, shuffle=True,
+    def dataloader(self, epoch_length=None):
+        """ Creates a torch dataloader out of a subclass of this class.
+
+        Args:
+            epoch_length (int, optional): The epoch length of the miner. If this length is not set or if it is larger than the dataset, 
+            then a dataloader for the entire dataset is returned. Otherwise, a dataloader for a subset of the dataset of epoch_length 
+            is returned. Defaults to None.
+
+        Returns:
+            torch.utils.data.dataloader.DataLoader: Pytorch dataloader.
+        """
+        # If epoch_length is set then we just need a slice of 
+        # the dataset we downloaded of length epoch_length. 
+        if epoch_length and epoch_length < len(self):
+            
+            # Set up upper bound of indices to fit the batch size we want. 
+            idx_bound = epoch_length * self.batch_size
+
+            # Collect enough random indices to batch together using batch_size into epoch_length batches
+            random_start = random.randint(0, len(self) - idx_bound)
+            indices = list(range(random_start, random_start + idx_bound))
+            
+            subset = Subset(self, indices)
+
+            # Set up dataloader
+            return DataLoader(subset,
+                            batch_size=self.batch_size,
+                            num_workers=self.config.dataloader.num_workers)
+        
+        # If epoch_length is not set or it is higher than the total size of the dataset,
+        #  then just shuffle dataset and return the whole thing.
+        return DataLoader(self,
+                            shuffle=True,
                             batch_size=self.batch_size,
                             num_workers=self.config.dataloader.num_workers)
     
@@ -154,3 +141,4 @@ class BittensorDataLoader():
     def __getitem__(self, idx):
         """returns the next batch from the dataset.
         """
+        pass
