+++ tests/integration_tests/subtensor_client.py
@@ -1,8 +1,6 @@
 import os
-import sys
 import time
 import pytest
-import asyncio
 import random
 import subprocess
