+++ tests/miners/text/test_gpt2.py
@@ -12,7 +12,8 @@ def test_run_gpt2_genesis():
     miner = Miner(
         epoch_length = 1,
         n_epochs = 1,
-        custom_dataset = './miners/TEXT/gpt2_genesis/genesis_dataset/'
+        batch_size_train=10,
+        block_size=1
     )
     miner.neuron.subtensor.connect = MagicMock(return_value = True)    
     miner.neuron.subtensor.subscribe = MagicMock(return_value = True)
